requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
click==7.1.2
cycler==0.10.0
Django==2.0.13
greenlet==1.0.0
importlib-metadata==4.0.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
mysql-connector==2.2.9
mysql-connector-python==8.0.24
mysqlclient==2.0.3
nltk==3.5
numpy==1.19.5
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
pandas==1.1.5
Pillow==8.1.2
protobuf==3.15.8
pyforest==1.0.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
regex==2021.3.17
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
sklearn==0.0
soupsieve==2.2
SQLAlchemy==1.4.9
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.59.0
typing-extensions==3.7.4.3
wordcloud==1.8.1
zipp==3.4.1
